CDAP-4693 Disable explore where possible.
